Uplift InferConst to rustc_type_ir
